refactor(cli): share path expansion helpers
