+++ packages/remnic-cli/src/bench-args.ts
@@ -1,4 +1,5 @@
 import path from "node:path";
+import { expandTilde } from "./path-utils.js";
 
 export type BenchAction = "help" | "list" | "run" | "check" | "report";
 
@@ -9,24 +10,6 @@ export interface ParsedBenchArgs {
   all: boolean;
   json: boolean;
   datasetDir?: string;
-}
-
-function resolveHomeDir(): string {
-  return process.env.HOME ?? process.env.USERPROFILE ?? "~";
-}
-
-function expandTilde(p: string): string {
-  if (p === "~" || p.startsWith("~/") || p.startsWith("~\\")) {
-    return resolveHomeDir() + p.slice(1);
-  }
-  const home = resolveHomeDir();
-  if (p === "$HOME" || p.startsWith("$HOME/") || p.startsWith("$HOME\\")) {
-    return home + p.slice(5);
-  }
-  if (p === "${HOME}" || p.startsWith("${HOME}/") || p.startsWith("${HOME}\\")) {
-    return home + p.slice(7);
-  }
-  return p;
 }
 
 export function readBenchOptionValue(argv: string[], flag: string): string | undefined {
